requirements.txt
jsonschema==2.5.1
PyYAML==3.11
requests==2.8.1
